- Fix wrong type for CBTBE_Punch_Target_Instability_Per_Attacker_Ton (was int32 instead of float)

+++ CBTBehaviorsEnhanced/CBTBehaviorsEnhanced/Patches/MechPatches.cs
@@ -68,7 +68,7 @@ namespace CBTBehaviorsEnhanced.Patches
             __instance.StatCollection.AddStatistic<int>(ModStats.PunchTargetDamageMod, 0);
             __instance.StatCollection.AddStatistic<float>(ModStats.PunchTargetDamageMulti, 1f);
 
-            __instance.StatCollection.AddStatistic<int>(ModStats.PunchTargetInstability, 0);
+            __instance.StatCollection.AddStatistic<float>(ModStats.PunchTargetInstability, 0);
             __instance.StatCollection.AddStatistic<int>(ModStats.PunchTargetInstabilityMod, 0);
             __instance.StatCollection.AddStatistic<float>(ModStats.PunchTargetInstabilityMulti, 1f);
